Cleanup unused kupo and pc-contracts-cli, query ogmios for network
