+++ toolkit/partner-chains-cli/src/main.rs
@@ -10,13 +10,11 @@ pub mod io;
 pub mod keystore;
 pub(crate) mod main_chain_follower;
 pub(crate) mod ogmios;
-pub(crate) mod pc_contracts_cli_resources;
 pub(crate) mod permissioned_candidates;
 mod prepare_configuration;
 pub mod register;
 pub(crate) mod select_utxo;
 mod setup_main_chain_state;
-pub(crate) mod smart_contracts;
 pub mod start_node;
 
 #[cfg(test)]
