+++ toolkit/partner-chains-cli/src/setup_main_chain_state/tests.rs
@@ -1,8 +1,7 @@
 use crate::config::config_fields;
 use crate::config::config_fields::CARDANO_PAYMENT_SIGNING_KEY_FILE;
 use crate::config::{CHAIN_CONFIG_FILE_PATH, RESOURCES_CONFIG_FILE_PATH};
-use crate::pc_contracts_cli_resources::default_ogmios_service_config;
-use crate::pc_contracts_cli_resources::tests::prompt_ogmios_configuration_io;
+use crate::ogmios::config::tests::{default_ogmios_service_config, prompt_ogmios_configuration_io};
 use crate::prepare_configuration::tests::{
 	prompt_and_save_to_existing_file, prompt_with_default_and_save_to_existing_file,
 };
